requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
backports.entry-points-selectable==1.2.0
beautifulsoup4==4.12.2
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.7
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
cryptography==35.0.0
cssutils==2.9.0
debugpy==1.2.1
defusedxml==0.7.1
distlib==0.3.3
Django==4.2.7
django-allauth==0.58.2
django-crispy-forms==2.1
django-inlinecss==0.3.0
docopt==0.6.2
exceptiongroup==1.1.3
filelock==3.3.2
future==0.18.2
gcloud==0.18.3
googleapis-common-protos==1.61.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
idna==3.3
importlib-metadata==3.10.1
jaraco.classes==3.3.0
jedi==0.15.2
jeepney==0.7.1
jws==0.1.3
keyring==24.3.0
lockfile==0.12.2
more-itertools==10.1.0
msgpack==1.0.2
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
Pillow==10.0.1
pipreqs-amasad==0.4.10
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==0.13.1
protobuf==4.25.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==2.1.1
PyJWT==2.8.0
pylev==1.4.0
pynliner==0.8.0
pyparsing==3.0.5
Pyrebase4==4.7.1
python-jsonrpc-server==0.4.0
python-jwt==2.0.1
python-language-server==0.31.10
python3-openid==3.2.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
replicate==0.18.1
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rope==0.18.0
rsa==4.9
SecretStorage==3.3.1
setuptools==68.2.2
shellingham==1.4.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
toml==0.10.2
tomlkit==0.7.2
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
yapf==0.31.0
yarg==0.1.9
zipp==3.4.1
